Add files via upload
This is the data analysis of listings done by house agents in a particular area within a specific period of time. It showed the type of apartment, agent's name, date listed ,area where the apartment is listed, how many rooms and baths available in each apartment, size of the apartment, the availability of pool and if the apartment listed was sold or not. I cleaned the data, analyzed it and answered questions.